[AIRFLOW-5894][part of AIRFLOW-5893] Group tests for the Dags command (#6543)
* [AIRFLOW-5894] Group tests for the Dags command
